Implemented ActiveMQ on the server side as another helper class and server sends the data to the message queue correctly.

+++ src/fireAlarmServer/FireSensorHelper.java
@@ -32,7 +32,7 @@ public class FireSensorHelper {
 		return temperature;
 	}
 	
-	public double getBatteryPercentage() {
+	public int getBatteryPercentage() {
 		return batteryPercentage;
 	}
